change maintenance page design
