+++ src/index.html
@@ -1,18 +1,18 @@
 <!doctype html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/x-icon" href="./assets/favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script
-      src="https://kit.fontawesome.com/d61bd32107.js"
-      crossorigin="anonymous"
-    ></script>
-    <script type="module" src="./index.tsx"></script>
-    <title>LÖVEBrew: Bundle Your Game</title>
-  </head>
+    <head>
+        <meta charset="UTF-8" />
+        <link rel="icon" type="image/x-icon" href="./assets/favicon.ico" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <script
+            src="https://kit.fontawesome.com/d61bd32107.js"
+            crossorigin="anonymous"
+        ></script>
+        <script type="module" src="./index.tsx"></script>
+        <title>LÖVEBrew: Bundle Your Game</title>
+    </head>
 
-  <body>
-    <div id="root"></div>
-  </body>
+    <body>
+        <div id="root"></div>
+    </body>
 </html>
