+++ src/styles/overlay.css
@@ -1,36 +1,73 @@
-.loader-containter {
-  position: absolute;
-  top: 1em;
-  right: 1em;
-}
-.overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  z-index: 1000;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  backdrop-filter: blur(5px);
+.maintenance-container {
+    position: fixed;
+    inset: 0;
+    background: whitesmoke;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    text-align: center;
+    padding: 2rem;
+    z-index: 9999;
+    font-family: "Nunito", sans-serif;
 }
 
-.overlay-content {
-  font-family: "Nunito";
-  text-align: center;
-  width: 90vw;
-  text-shadow: 0 0 0.5rem rgba(0, 0, 0, 0.85);
+.maintenance-logo {
+    width: 128px;
+    height: 128px;
+    object-fit: contain;
+    margin: 0 auto 1.5rem auto;
+    display: block;
+    opacity: 0.9;
+    animation: potionAppear 1s ease-out 0s 1;
 }
 
-.overlay-content h1 {
-  font-size: 2rem;
-  color: #fff;
+.maintenance-box {
+    max-width: 500px;
+    padding: 2.5rem;
+    border-radius: 12px;
 }
 
-.overlay-content p {
-  font-size: 1.2rem;
-  color: #fff;
+.maintenance-title {
+    font-size: 2rem;
+    font-weight: 700;
+    color: #111;
+    margin-bottom: 1rem;
+}
+
+.maintenance-text {
+    font-size: 1rem;
+    line-height: 1.6;
+    margin-bottom: 1.5rem;
+    color: #444;
+}
+
+.maintenance-button {
+    display: inline-block;
+    background: #0070f3;
+    color: white;
+    padding: 0.75rem 1.5rem;
+    border-radius: 4px;
+    text-decoration: none;
+    font-weight: 600;
+    transition: background 0.2s ease-in-out;
+}
+
+.maintenance-button:hover {
+    background: #005ac1;
+}
+
+/* Optional subtle fade-in animation */
+@keyframes fadeUp {
+    from {
+        opacity: 0;
+        transform: translateY(10px);
+    }
+    to {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
+
+.maintenance-box {
+    animation: fadeUp 0.4s ease-out;
 }
